added and renamed getters
